+++ frontend/javascripts/test/puppeteer/dataset_rendering.screenshot.js
@@ -92,7 +92,14 @@ const datasetConfigOverrides: { [key: string]: DatasetConfiguration } = {
     fourBit: false,
     interpolation: true,
     layers: {
-      color: { color: [255, 255, 255], contrast: 1, brightness: 0, alpha: 100, isDisabled: false },
+      color: {
+        color: [255, 255, 255],
+        contrast: 1,
+        brightness: 0,
+        alpha: 100,
+        intensityRange: [0, 255],
+        isDisabled: false,
+      },
     },
     quality: 0,
     segmentationOpacity: 0,